Auto stash before merge of "master" and "origin/master"
for Authetication

+++ app/Http/Controllers/AddFlowers_to_session_BQT.php
@@ -42,12 +42,12 @@ class AddFlowers_to_session_BQT extends Controller
    public function store(Request $request)
     {
         //
-        /*if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
             Cart::instance('OrderedBqt_Flowers')->destroy();
         }
-        else{     */
+        else{
              Cart::instance('OrderedBqt_Flowers')->count();
 
             $Flower_ID = $request->BqtFlwrID_Field;
@@ -190,12 +190,12 @@ class AddFlowers_to_session_BQT extends Controller
     public function update(Request $request, $id)
     {
 
-       /* if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
             Cart::instance('OrderedBqt_Flowers')->destroy();
         }
-        else{  */
+        else{
           $newQty = $request->QuantityField;
           $order_ID = $request->ID_Field;
           $descision = $request->Decision_Field;
