+++ app/Http/Controllers/AddFlowers_to_session_QuickBQT.php
@@ -43,12 +43,12 @@ class AddFlowers_to_session_QuickBQT extends Controller
     public function store(Request $request)
     {
       //
-      /*if(auth::check() == false){
+      if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
           Cart::instance('OrderedBqt_Flowers')->destroy();
       }
-      else{     */
+      else{     
 
       //dd(Cart::instance('overallFLowers')->content());
 
@@ -262,12 +262,12 @@ class AddFlowers_to_session_QuickBQT extends Controller
      */
     public function update(Request $request, $id)
     {
-      /* if(auth::check() == false){
+       if(auth::guard('admins')->check() == false){
            Session::put('loginSession','fail');
            return redirect() -> route('adminsignin');
            Cart::instance('OrderedBqt_Flowers')->destroy();
        }
-       else{  */
+       else{  
          $newQty = $request->QuantityField;
          $order_ID = $request->ID_Field;
          $descision = $request->Decision_Field;
