+++ app/Http/Controllers/OrderManagementController.php
@@ -331,11 +331,11 @@ foreach($debtDetails as $debtDetails){
 
     public function DeleteFlower_per_Order($flower_ID)
 	{
-        /*if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-	   else{*/
+	   else{
 	        $AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
 
 			foreach(Cart::instance('Ordered_Flowers')->content() as $row){
@@ -352,11 +352,11 @@ foreach($debtDetails as $debtDetails){
 	}
     public function DeleteFlower_per_Bqt_Order($flower_ID,$order_ID)
 	{
-		/*if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 			echo $flower_ID;
 			foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
 				if($row->id == $flower_ID){
@@ -371,11 +371,11 @@ foreach($debtDetails as $debtDetails){
 
 	public function DeleteFlower_per_Bqt_SessionOrder($flower_ID)
 	{
-        /*if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 			echo $flower_ID;
 			foreach(Cart::instance('OrderedBqt_Flowers')->content() as $row){
 				if($row->id == $flower_ID){
@@ -393,7 +393,7 @@ foreach($debtDetails as $debtDetails){
 
     public function DeleteAcessory_per_Bqt_Order($Acessory_ID,$order_ID)
 	{
-        /*if(auth::check() == false){
+        /*if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
@@ -412,7 +412,7 @@ foreach($debtDetails as $debtDetails){
 
     public function DeleteAcessory_per_SessionBqt_Order($Acessory_ID)
 	{
-    if(auth::check() == false){
+    if(auth::guard('admins')->check() == false){
         Session::put('loginSession','fail');
         return redirect() -> route('adminsignin');
     }
@@ -431,11 +431,11 @@ foreach($debtDetails as $debtDetails){
 
 	public function Cancel_and_ClearFlower_per_Bqt_Order($order_ID)
 	{
-		/*if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 			$AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
 
 			Cart::instance('OrderedBqt_Flowers')->destroy();
@@ -450,10 +450,10 @@ foreach($debtDetails as $debtDetails){
 
 	public function Cancel_and_Clear_Bqt_Order()
 	{
-		/*if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');//
-        else{*/
+        else{
 			$AvailableFlowers = DB::select('call wonderbloomdb2.Viewing_Flowers_With_UpdatedPrice()');
 
 			Cart::instance('OrderedBqt_Flowers')->destroy();
@@ -466,7 +466,7 @@ foreach($debtDetails as $debtDetails){
 
 	public function Cancel_and_Clear_BqtSession_Order()
 	{
-    if(auth::check() == false){
+    if(auth::guard('admins')->check() == false){
         Session::put('loginSession','fail');
         return redirect() -> route('adminsignin');
     }
@@ -487,7 +487,7 @@ foreach($debtDetails as $debtDetails){
 
     public function Clear_Cart()
     {
-     if(auth::check() == false){
+     if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
       }
@@ -506,7 +506,7 @@ foreach($debtDetails as $debtDetails){
 
     public function Clear_Bouquet()
     {
-      if(auth::check() == false){
+      if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
       }
@@ -523,7 +523,7 @@ foreach($debtDetails as $debtDetails){
 
 	public function saveCustomized_Bqt($order_ID)
 	{
-    if(auth::check() == false){
+    if(auth::guard('admins')->check() == false){
           Session::put('loginSession','fail');
           return redirect() -> route('adminsignin');
       }
@@ -588,7 +588,7 @@ foreach($debtDetails as $debtDetails){
 
 
 	public function saveNewCustomized_Bqt(){
-        if(auth::check() == false){
+        if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
@@ -659,11 +659,11 @@ foreach($debtDetails as $debtDetails){
 
 	public function ConfrimOrder()
 	{//
-		/*if(auth::check() == false){
+		if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
-        else{*/
+        else{
 
 	          $cities = DB::table('cities')
 	          ->select('*')
@@ -681,7 +681,7 @@ foreach($debtDetails as $debtDetails){
 
 	public function return_to_CreationOfOrder()
 	{//
-       if(auth::check() == false){
+       if(auth::guard('admins')->check() == false){
             Session::put('loginSession','fail');
             return redirect() -> route('adminsignin');
         }
